Moved occupations to a separate slice
