+++ src/api.ts
@@ -1,6 +1,6 @@
 import * as Mocks from "./mocks";
 import * as HotelSlice from "./redux/hotelSlice";
-import * as TableSlice from "./redux/tableSlice";
+import * as OccupationsSlice from "./redux/occupationsSlice";
 
 export function fetchHotelDataAsync(): Promise<{ data: HotelSlice.HotelData }> {
   return new Promise((resolve) => {
@@ -8,7 +8,7 @@ export function fetchHotelDataAsync(): Promise<{ data: HotelSlice.HotelData }> {
   });
 }
 
-export function fetchTilesAsync(from: string, to: string): Promise<{ data: TableSlice.TileData[] }> {
+export function fetchTilesAsync(from: string, to: string): Promise<{ data: OccupationsSlice.TileData[] }> {
   console.log(`fetching tiles from ${from} to ${to}...`);
   return new Promise((resolve) => {
     setTimeout(() => resolve({ data: Mocks.tiles }), 500);
