+++ src/components/table-container/TableCell.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { hot } from "react-hot-loader";
 
 import { useAppSelector } from "../../redux/hooks";
-import * as TableSlice from "../../redux/tableSlice";
+import * as OccupationsSlice from "../../redux/occupationsSlice";
 
 import Tile from "./Tile";
 
@@ -31,10 +31,9 @@ function TableCell(props: Props): JSX.Element {
   return <div className="table-cell">{tile}</div>;
 }
 
-function useTileDataAt(x: string, y: number): TableSlice.TileData | undefined {
+function useTileDataAt(x: string, y: number): OccupationsSlice.TileData | undefined {
   return useAppSelector(state => {
-    const occupations = state.table.occupations;
-    const occupationsForRoom = occupations[y];
+    const occupationsForRoom = state.occupations[y];
     return (occupationsForRoom === undefined) ? undefined : occupationsForRoom[x];
   });
 }
